requirements.txt
kafka-python==2.0.2
six==1.16.0
matplotlib==3.9.0
numpy==1.26.4
python-dotenv==1.0.1
